+++ src/Microsoft.ML.Data/StaticPipe/StaticPipeUtils.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using Microsoft.ML.Core.Data;
 using Microsoft.ML.Runtime;
 using Microsoft.ML.Runtime.Data;
+using Microsoft.ML.Transforms;
 
 namespace Microsoft.ML.StaticPipe.Runtime
 {
