+++ src/Microsoft.ML.Data/StaticPipe/TrainerEstimatorReconciler.cs
@@ -2,12 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections.Generic;
-using System.Linq;
 using Microsoft.ML.Core.Data;
 using Microsoft.ML.Runtime;
 using Microsoft.ML.Runtime.Data;
 using Microsoft.ML.Runtime.Internal.Utilities;
+using Microsoft.ML.Transforms;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Microsoft.ML.StaticPipe.Runtime
 {
