+++ test/BaselineOutput/Common/EntryPoints/core_ep-list.tsv
@@ -82,7 +82,7 @@ Transforms.CategoricalHashOneHotVectorizer	Converts the categorical value into a
 Transforms.CategoricalOneHotVectorizer	Converts the categorical value into an indicator array by building a dictionary of categories based on the data and using the id in the dictionary as the index in the array.	Microsoft.ML.Runtime.Data.Categorical	CatTransformDict	Microsoft.ML.Runtime.Data.CategoricalTransform+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+TransformOutput
 Transforms.CharacterTokenizer	Character-oriented tokenizer where text is considered a sequence of characters.	Microsoft.ML.Runtime.Transforms.TextAnalytics	CharTokenize	Microsoft.ML.Runtime.TextAnalytics.CharTokenizeTransform+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+TransformOutput
 Transforms.ColumnConcatenator	Concatenates one or more columns of the same item type.	Microsoft.ML.Runtime.EntryPoints.SchemaManipulation	ConcatColumns	Microsoft.ML.Runtime.Data.ConcatTransform+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+TransformOutput
-Transforms.ColumnCopier	Duplicates columns from the dataset	Microsoft.ML.Runtime.EntryPoints.SchemaManipulation	CopyColumns	Microsoft.ML.Runtime.Data.CopyColumnsTransform+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+TransformOutput
+Transforms.ColumnCopier	Duplicates columns from the dataset	Microsoft.ML.Runtime.EntryPoints.SchemaManipulation	CopyColumns	Microsoft.ML.Transforms.CopyColumnsTransform+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+TransformOutput
 Transforms.ColumnDropper	Drops columns from the dataset	Microsoft.ML.Runtime.EntryPoints.SchemaManipulation	DropColumns	Microsoft.ML.Runtime.Data.DropColumnsTransform+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+TransformOutput
 Transforms.ColumnSelector	Selects a set of columns, dropping all others	Microsoft.ML.Runtime.EntryPoints.SchemaManipulation	SelectColumns	Microsoft.ML.Runtime.Data.DropColumnsTransform+KeepArguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+TransformOutput
 Transforms.ColumnTypeConverter	Converts a column to a different type, using standard conversions.	Microsoft.ML.Runtime.Data.TypeConversion	Convert	Microsoft.ML.Runtime.Data.ConvertTransform+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+TransformOutput
